Define initial record types for the consumer offset topic
The goal is to define initial record types for storing group metadata. The aim is roughly match the data represented in the RPCs and the KIP-848 records. There is an extra record for the topology that is called when StreamsInitialize is called.

See https://github.com/lucasbru/kafka/pull/14

target assignment and topology

current assignment

member metadata

names and version

initial copy