Fix failing to recycle task blocks active tasks
If a "pending" StandbyTask is assigned to an instance as an Active, we
try to recycle our existing StandbyTask into the Active. However, if
this fails, we were silently discarding this broken StandbyTask without
ever actually creating an Active.

Now, when we fail to recycle the StandbyTask, we close it dirty (to
prevent resource leaks and clean up any bad state) and we no longer
remove it from the set of active Tasks that need to be created. This
ensures that an Active Task will be created (from scratch) later in the
assignment process.