Resolve conflict from 11/25 trunk rebase - Streams reconciliation logic
This PR implements the new reconciliation logic for Kafka Streams. In
the POC, we used the reconciliation logic for topic partitions on the
broker side, developed as part of KIP-848. This worked for active tasks,
but does not take standby tasks or warm-up tasks into account.
The new reconciliation logic also takes standby and warm up tasks into
account. Inside the streams group object, we keep track of the process
IDs that own a task, and in which role. This replaces task epochs from
KIP-848. Using the process IDs, we can check the assignment invariants:

* A. A process can own a task only once in any role. For example, if
processX already owns a task as a standby, it cannot own it as an active
as well.

* B. For active tasks, since there can never two instances owning the
same task in an active role, the invariant is different: An active task
can be assigned only if no other member owns it (as an active task).

The reconciliation logic will make sure that these invariants are
followed. For this, it uses the same states for the member as KIP-848,
just with different state transitions:

A member is in state UNREVOKED whenever we are waiting for the member to
confirm revocation of at least one task. We expect the member to revoke
all tasks that are not in its target assignment before transitioning to
the next epoch. Once the member has revoked all tasks that are not in
its target assignment, it transitions to the new epoch and gets new
tasks assigned. Once all tasks are assigned, we transition to STABLE,
but it may be that some tasks cannot be assigned yet, because of the
above invariants:

* A. For any task, another instance of the same process may own the task
in any role (active/standby/warmup), and we are still awaiting
revocation.

* B. For an active task, an instance of any process may own the same
active task, and we are still awaiting revocation.

As long as tasks cannot be assigned, we remain in state UNRELEASED.

The testing required some changes, so that I could parameterize the test
on the task role which is causing state transitions. There are some
utitlity changes / clean-ups to construct assignments.