Don't special case failure to recycle pending task
When a pending task fails to be recycled, I was capturing this and
attempting to create a new task in its place.

However, this was pointless, because a failure to recycle a task causes
an exception to be thrown for that task, so creating a new task in its
place would have no effect.

Removing this special case simplifies things a bit, as it now works like
the rest of the assignment.